Changed Get-Item command to find parent directory
I added a -Force flag to the Get-Item command so that you can use
posh-git in hidden directories. Without the -Force flag, PowerShell will
error when you call Get-Item on a hidden directory, even if you're in it.

I mainly added this so that I could use posh-git on my vimfiles
directory which is hidden at ~/vimfiles.